+++ assets/javascripts/kitten/components/form/text-input-with-button.js
@@ -46,6 +46,7 @@ export class TextInputWithButton extends Component {
         />
         <button
           { ...buttonProps }
+          disabled={ disabled }
           className={ textInputWithButtonButtonClassName }
           type="button"
         >