+++ assets/javascripts/kitten/components/carousel/carousel-inner.js
@@ -1,5 +1,6 @@
 import React from 'react'
-import Radium, { Style } from 'radium'
+import { Style } from 'radium'
+import { ConfiguredRadium } from 'kitten/config/configured-radium'
 import ResizeObserver from 'resize-observer-polyfill'
 
 if (typeof window !== 'undefined') {
@@ -235,4 +236,4 @@ const styles = {
   },
 }
 
-export const CarouselInner = Radium(CarouselInnerBase)
+export const CarouselInner = ConfiguredRadium(CarouselInnerBase)