+++ CHANGELOG.md
@@ -4,6 +4,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Feature: Add `ConfiguredRadium` HOC to use Radium with custom configuration.
 - Fix: Update Radium version and add a custom plugin to fix `prefixer` data.
 - Fix: `Carousel` prerender.