+++ CHANGELOG.md
@@ -7,6 +7,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Feature: Add `ConfiguredRadium` HOC to use Radium with custom configuration.
 - Fix: Update Radium version and add a custom plugin to fix `prefixer` data.
 - Fix: Fix `Carousel` prerender.
+- Fix: Update styles on `Tour`, `Popover` and `Modal` components.
+- Feature: Add `k-ButtonIcon--cross--tiny` class.
 
 ## [19.0.0] - 2018-01-16