+++ CHANGELOG.md
@@ -4,6 +4,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Feature: Update styles `disabled` prop on `Button` and `ButtonIcon`.
+
 ## [19.1.0] - 2018-01-22
 
 Features: