+++ CHANGELOG.md
@@ -5,6 +5,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [unreleased]
 
 - Feature: Update styles `disabled` prop on `Button` and `ButtonIcon`.
+- Fix: Add `assets/javascripts/kitten/config` folder to published package.
 
 ## [19.1.0] - 2018-01-22