+++ CHANGELOG.md
@@ -5,6 +5,19 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [unreleased]
 
 - Fix: Update colors to `TextInputWithButton` component.
+- Feature: Update styles `disabled` prop on `Button` and `ButtonIcon`.
+- Fix: Add `assets/javascripts/kitten/config` folder to published package.
+
+## [19.1.0] - 2018-01-22
+
+Features:
+- Add `ConfiguredRadium` HOC to use Radium with custom configuration.
+- Add `k-ButtonIcon--cross--tiny` class.
+
+Fixes:
+- Update Radium version and add a custom plugin to fix `prefixer` data.
+- Fix `Carousel` prerender.
+- Update styles on `Tour`, `Popover` and `Modal` components.
 
 ## [19.0.0] - 2018-01-16